use int for server history indexes rather than long

+++ src/cpp/r/include/r/session/RConsoleHistory.hpp
@@ -47,7 +47,7 @@ private:
    // COPYING: boost::noncopyable
       
 public:   
-   void setCapacity(long capacity);
+   void setCapacity(int capacity);
 
    void setRemoveDuplicates(bool removeDuplicates);
    
@@ -56,15 +56,15 @@ public:
    const_iterator begin() const { return historyBuffer_.begin(); }
    const_iterator end() const { return historyBuffer_.end(); }
    
-   long size() const
+   int size() const
    {
       return historyBuffer_.size();
    }
 
    void clear();
 
-   void subset(long beginIndex, // inclusive
-               long endIndex,   // exclusive,
+   void subset(int beginIndex, // inclusive
+               int endIndex,   // exclusive,
                std::vector<std::string>* pEntries) const;
 
    void asJson(core::json::Array* pHistoryArray) const;
